[BPF] Convert BlockAddress to ConstantArray
In previous patch, for Example 2, we have

       0:       18 02 00 00 50 00 00 00 00 00 00 00 00 00 00 00 r2 = 0x50 ll
                0000000000000000:  R_BPF_64_64  .text
       2:       16 01 02 00 00 00 00 00 if w1 == 0x0 goto +0x2 <bar+0x28>
       3:       18 02 00 00 40 00 00 00 00 00 00 00 00 00 00 00 r2 = 0x40 ll
                0000000000000018:  R_BPF_64_64  .text
       5:       7b 2a f8 ff 00 00 00 00 *(u64 *)(r10 - 0x8) = r2
       6:       79 a1 f8 ff 00 00 00 00 r1 = *(u64 *)(r10 - 0x8)
       7:       0d 01 00 00 00 00 00 00 gotox r1
       8:       b4 00 00 00 03 00 00 00 w0 = 0x3
       9:       05 00 01 00 00 00 00 00 goto +0x1 <bar+0x58>
      10:       b4 00 00 00 02 00 00 00 w0 = 0x2
      11:       95 00 00 00 00 00 00 00 exit

For this case, there is no jump table so it would be hard to track offset
during verification. So practically we need to create two jump tables
for '&&l1' and '&&l2' respectively.

This patch converts all 'blockaddress' references to global variables
where each global variable is an array with size 1. For example,
the original insn is:
    reg = LD_imm64 <blockaddress>

The global variable will be ba2gv[1] and the single element will be
    ba2gv[0] = <blockaddress>
With this, the new insn sequence (replacing 'reg = LD_imm64 <blockaddress>')
will be:
    reg = LD_imm64 ba2gv
    reg = *(u64 *)(reg + 0)
This way, we can put global variable in .jumptables section.
The following is the new obj code:

       0:       18 02 00 00 00 00 00 00 00 00 00 00 00 00 00 00 r2 = 0x0 ll
                0000000000000000:  R_BPF_64_64  .rodata
       2:       79 22 00 00 00 00 00 00 r2 = *(u64 *)(r2 + 0x0)
       3:       16 01 03 00 00 00 00 00 if w1 == 0x0 goto +0x3 <bar+0x38>
       4:       18 02 00 00 08 00 00 00 00 00 00 00 00 00 00 00 r2 = 0x8 ll
                0000000000000020:  R_BPF_64_64  .rodata
       6:       79 22 00 00 00 00 00 00 r2 = *(u64 *)(r2 + 0x0)
       7:       7b 2a f8 ff 00 00 00 00 *(u64 *)(r10 - 0x8) = r2
       8:       79 a1 f8 ff 00 00 00 00 r1 = *(u64 *)(r10 - 0x8)
       9:       0d 01 00 00 00 00 00 00 gotox r1
      10:       b4 00 00 00 03 00 00 00 w0 = 0x3
      11:       05 00 01 00 00 00 00 00 goto +0x1 <bar+0x68>
      12:       b4 00 00 00 02 00 00 00 w0 = 0x2
      13:       95 00 00 00 00 00 00 00 exit

With this patch, the MO_BlockAddress processing in BPFMCInstLower.cpp is
not needed anymore since All BlockAddress references have been replaced
by global variables.