[LLVM] Emit dwarf data for changed-signature and new functions
Add a new pass EmitChangedFuncDebugInfo which will add dwarf for
additional functions including functions with signature change
and new functions.

The previous approach in [1] tries to add debuginfo for those
optimization passes which cause signature changes. Based on
discussion in [1], it is preferred to have a specific pass to
add debuginfo and later on dwarf generation can include those
new debuginfo.

The following is an example:
Source:
  $ cat test.c
  struct t { int a; };
  char *tar(struct t *a, struct t *d);
  __attribute__((noinline)) static char * foo(struct t *a, struct t *d, int b)
  {
    return tar(a, d);
  }
  char *bar(struct t *a, struct t *d)
  {
    return foo(a, d, 1);
  }
Compiled and dump dwarf with:
  clang -O2 -c -g test.c
  llvm-dwarfdump test.o

0x0000005c:   DW_TAG_subprogram
...
0x000000ba:     NULL

There are some restrictions in the current implementation:
  - Only C language is supported
  - BPF target is excluded as one of main goals for this pull request
    is to generate proper vmlinux BTF for arch's like x86_64/arm64 etc.
  - Function must not be a intrinsic, decl only, return value size more
    than arch register size and func with variable arguments.

I have tested this patch set by building latest bpf-next linux kernel.
For no-lto case:
  65341 original number of functions
  1085  new functions with this patch
For thin-lto case:
  65595 original number of functions
  2492  new functions with this patch

  [1] https://github.com/llvm/llvm-project/pull/127855